first commit + state machine
