+++ Client-PC/Network-Stack/go.mod
@@ -1,0 +1,7 @@
+module networkstack
+
+go 1.25.0
+
+require github.com/0xcafed00d/joystick v1.0.1
+
+require golang.org/x/sys v0.41.0 // indirect
