+++ Server-Pi/Network-Stack/go.mod
@@ -1,0 +1,10 @@
+module networkstack
+
+go 1.25.0
+
+require go.bug.st/serial v1.6.4
+
+require (
+	github.com/creack/goselect v0.1.2 // indirect
+	golang.org/x/sys v0.19.0 // indirect
+)
