+++ Server-Pi/Neural-Network/Gazebo/worlds/lunafield_heightmap.png
