+++ Server-Pi/Neural-Network/Gazebo/worlds/my_world.sdf
