+++ Server-Pi/Neural-Network/Gazebo/worlds/world_with_dummy_rover.sdf
